chore: convert inboxproductaanvraag backend code to Kotlin
